ISSUES-4006 add factor with DateTime type
