+++ src/DataTypes/DataTypeDateTime64.cpp
@@ -201,65 +201,4 @@ bool DataTypeDateTime64::equals(const IDataType & rhs) const
     return false;
 }
 
-namespace ErrorCodes
-{
-    extern const int NUMBER_OF_ARGUMENTS_DOESNT_MATCH;
-    extern const int ILLEGAL_TYPE_OF_ARGUMENT;
-}
-
-enum class ArgumentKind
-{
-    Optional,
-    Mandatory
-};
-
-template <typename T, ArgumentKind Kind>
-std::conditional_t<Kind == ArgumentKind::Optional, std::optional<T>, T>
-getArgument(const ASTPtr & arguments, size_t argument_index, const char * argument_name, const std::string context_data_type_name)
-{
-    using NearestResultType = NearestFieldType<T>;
-    const auto field_type = Field::TypeToEnum<NearestResultType>::value;
-    const ASTLiteral * argument = nullptr;
-
-    auto exception_message = [=](const String & message)
-    {
-        return std::string("Parameter #") + std::to_string(argument_index) + " '"
-                + argument_name + "' for " + context_data_type_name
-                + message
-                + ", expected: " + Field::Types::toString(field_type) + " literal.";
-    };
-
-    if (!arguments || arguments->children.size() <= argument_index
-        || !(argument = arguments->children[argument_index]->as<ASTLiteral>()))
-    {
-        if constexpr (Kind == ArgumentKind::Optional)
-            return {};
-        else
-            throw Exception(exception_message(" is missing"),
-                            ErrorCodes::NUMBER_OF_ARGUMENTS_DOESNT_MATCH);
-    }
-
-    if (argument->value.getType() != field_type)
-        throw Exception(exception_message(String(" has wrong type: ") + argument->value.getTypeName()),
-                        ErrorCodes::ILLEGAL_TYPE_OF_ARGUMENT);
-
-    return argument->value.get<NearestResultType>();
-}
-
-static DataTypePtr create64(const ASTPtr & arguments)
-{
-    if (!arguments || arguments->size() == 0)
-        return std::make_shared<DataTypeDateTime64>(DataTypeDateTime64::default_scale);
-
-    const auto scale = getArgument<UInt64, ArgumentKind::Optional>(arguments, 0, "scale", "DateType64");
-    const auto timezone = getArgument<String, ArgumentKind::Optional>(arguments, !!scale, "timezone", "DateType64");
-
-    return std::make_shared<DataTypeDateTime64>(scale.value_or(DataTypeDateTime64::default_scale), timezone.value_or(String{}));
-}
-
-void registerDataTypeDateTime64(DataTypeFactory & factory)
-{
-    factory.registerDataType("DateTime64", create64, DataTypeFactory::CaseInsensitive);
-}
-
 }
