+++ tests/queries/0_stateless/01442_date_time_with_params.reference
@@ -1,0 +1,4 @@
+2020-01-01 00:00:00	DateTime	2020-01-01 00:01:00	DateTime	2020-01-01 00:02:00.11	DateTime64(2)	2020-01-01 00:03:00	DateTime(\'Europe/Moscow\')	2020-01-01 00:04:00.220	DateTime64(3, \'Europe/Moscow\')	2020-01-01 00:05:00	DateTime	2020-01-01 00:06:00	DateTime(\'Europe/Moscow\')
+2020-01-01 00:00:00	DateTime	2020-01-01 00:02:00.11	DateTime64(2)	2020-01-01 00:03:00	DateTime(\'Europe/Moscow\')	2020-01-01 00:04:00.220	DateTime64(3, \'Europe/Moscow\')
+2020-01-01 00:00:00	DateTime	2020-01-01 00:02:00.11	DateTime64(2)	2020-01-01 00:03:00	DateTime(\'Europe/Moscow\')	2020-01-01 00:04:00.220	DateTime64(3, \'Europe/Moscow\')
+2020-01-01 00:00:00	DateTime
